New Monster: Dread Lich
(Name is TBD) To keep some more iconic Lich spells around in this single
book world, the Dread Lich is the lich of demon summoning, with a book
consisting of Haste, Summon Greater Demon, Paralyze, and Invisibility.

Dread liches have the same stats as ancient liches, and appear in the
same places. In vaults they either increased the total lich weight
(where the vault seemed weak), or split weight with ancient liches,
except in those cases when the vault designer seemd to specifically be
using ancient liches in their summoner or their conjuror role.